Use next dynamic and move file
