+++ apps/web/src/app/[locale]/layout.tsx
@@ -4,11 +4,21 @@ import "../../style.css";
 import languages from "@rallly/languages";
 import { Toaster } from "@rallly/ui/toaster";
 import { Viewport } from "next";
+import dynamic from "next/dynamic";
 import { Inter } from "next/font/google";
 import React from "react";
 
-import { TimeZoneChangeDetector } from "@/app/[locale]/(admin)/timezone-change-detector";
 import { Providers } from "@/app/providers";
+
+const TimeZoneChangeDetector = dynamic(
+  () =>
+    import("./timezone-change-detector").then(
+      (mod) => mod.TimeZoneChangeDetector,
+    ),
+  {
+    ssr: false,
+  },
+);
 
 const inter = Inter({
   subsets: ["latin"],
